QueryPushpinStat/server/requirements.txt
asgiref==2.3.2
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.7.1
Automat==0.7.0
certifi==2018.4.16
channels==2.1.2
chardet==3.0.4
constantly==15.1.0
daphne==2.2.1
Django==2.1
django-dotenv==1.4.2
django-eventstream==2.1.0
django-grip==1.8.0
djangorestframework==3.8.2
gripcontrol==3.2.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
pubcontrol==2.4.2
PyHamcrest==1.9.0
PyJWT==1.6.4
python-dotenv==0.9.1
pytz==2018.5
pyzmq==17.1.0
requests==2.19.1
six==1.11.0
tnetstring3==0.3.1
Twisted==18.7.0
txaio==18.7.1
urllib3==1.23
Werkzeug==0.14.1
zmq==0.0.0
zope.interface==4.5.0